refactor portfolio components structure
