+++ src/components/sections/allPortfolio.js
@@ -4,35 +4,35 @@ import { useStaticQuery, graphql } from 'gatsby';
 import { ProjectCard } from '@components';
 import BubbleLink from '@styles/bubbleLink';
 import LandingSection from '@styles/landingSection';
-import { projectTitleChoices } from '@config';
 
 const Projects = ({ hideViewAllBtn }) => {
   const data = useStaticQuery(graphql`
-  query {
-    projects: allMarkdownRemark(
-  filter: {
-    fileAbsolutePath: {regex: "/portfolio/" }
-  frontmatter: {showInProjects: {ne: false } }
+    query {
+      projects: allMarkdownRemark(
+        filter: { 
+          fileAbsolutePath: { regex: "/portfolio/" }
+          frontmatter: { featureSection: { regex: "/(landing)|(portfolio)/"  } }
         }
-  sort: {fields: [frontmatter___date], order: DESC }
-  ) {
-    edges {
-    node {
-    frontmatter {
-    title
+        sort: { fields: [frontmatter___date], order: DESC }
+      ) {
+        edges {
+          node {
+            frontmatter {
+              date
+              description
+              title
               tech
-  github
-  external
-  description
-  thumbnail {
-    childImageSharp {
-    fluid(quality: 100, maxWidth: 1000) {
-    ...GatsbyImageSharpFluid
-  }
+              github
+              external
+              thumbnail {
+                childImageSharp {
+                  fluid(quality: 100, maxWidth: 1000) {
+                    ...GatsbyImageSharpFluid
+                  }
                 }
               }
-  slug
-  isFeatured
+              slug
+              featureSection
             }
           }
         }
@@ -40,17 +40,15 @@ const Projects = ({ hideViewAllBtn }) => {
     }
   `);
 
-  const projectsToShow = data.projects.edges.filter(({ node }) => projectTitleChoices.includes(node.frontmatter.title) && node.frontmatter.tech);
-
   return (
     <LandingSection id="portfolio">
-      <div className="text-center mx-auto max-w-4xl py-20">
-        <h2 className="text-title-text mt-6">Portfolio</h2>
+      <div className="text-center mx-auto max-w-4xl">
+        <h2 className="text-title-text mt-0">Portfolio</h2>
 
         {!hideViewAllBtn && <BubbleLink linkPath='/portfolio' isOutlined color="link">View All</BubbleLink>}
 
-        {projectsToShow &&
-          projectsToShow.map(({ node }) => (
+        {data.projects.edges &&
+          data.projects.edges.map(({ node }) => (
             <React.Fragment key={node.frontmatter.slug}>
               {node.frontmatter.tech && (
                 <ProjectCard node={node} />
