+++ src/components/sections/landingPortfolio.js
@@ -5,46 +5,40 @@ import { ProjectCard } from '@components';
 import BubbleLink from '@styles/bubbleLink';
 import LandingSection from '@styles/landingSection';
 
-// const MAX_PROJECTS_SHOWN = 3;
-
 const Projects = ({ hideViewAllBtn }) => {
   const data = useStaticQuery(graphql`
-  query {
-    projects: allMarkdownRemark(
-  filter: {
-    fileAbsolutePath: {regex: "/portfolio/" }
-  frontmatter: {showInProjects: {ne: false } }
+    query {
+      projects: allMarkdownRemark(
+        filter: { 
+          fileAbsolutePath: { regex: "/portfolio/" }
+          frontmatter: { featureSection: { eq: "landing" } }
         }
-  sort: {fields: [frontmatter___date], order: DESC }
-  ) {
-    edges {
-    node {
-    frontmatter {
-    title
+        sort: { fields: [frontmatter___date], order: DESC }
+      ) {
+        edges {
+          node {
+            frontmatter {
+              date
+              description
+              title
               tech
-  github
-  external
-  description
-  thumbnail {
-    childImageSharp {
-    fluid(quality: 100, maxWidth: 1000) {
-    ...GatsbyImageSharpFluid
-  }
+              github
+              external
+              thumbnail {
+                childImageSharp {
+                  fluid(quality: 100, maxWidth: 1000) {
+                    ...GatsbyImageSharpFluid
+                  }
                 }
               }
-  slug
-  isFeatured
+              slug
+              featureSection
             }
           }
         }
       }
     }
   `);
-
-  // const projectsToShow = data.projects.edges.filter(({ node }) => node.frontmatter.isFeatured && node.frontmatter.tech).slice(0, MAX_PROJECTS_SHOWN);
-  // Manually show SquidShop, QuickCart, and Woofer
-  const titleChoices = ['SquidShop Ecommerce', 'QuickCart', 'Atelier API'];
-  const projectsToShow = data.projects.edges.filter(({ node }) => titleChoices.includes(node.frontmatter.title) && node.frontmatter.tech);
 
   return (
     <LandingSection id="portfolio">
@@ -53,8 +47,8 @@ const Projects = ({ hideViewAllBtn }) => {
 
         {!hideViewAllBtn && <BubbleLink linkPath='/portfolio' isOutlined color="link">View All</BubbleLink>}
 
-        {projectsToShow &&
-          projectsToShow.map(({ node }) => (
+        {data.projects.edges &&
+          data.projects.edges.map(({ node }) => (
             <React.Fragment key={node.frontmatter.slug}>
               {node.frontmatter.tech && (
                 <ProjectCard node={node} />
