+++ src/pages/portfolio/index.js
@@ -1,20 +1,18 @@
 import React from 'react';
-import { graphql } from 'gatsby';
 import PropTypes from 'prop-types';
 import { Helmet } from 'react-helmet';
 import { Layout } from '@components';
-import PortfolioSection from '@components/sections/portfolio';
-import MoreProjectsSection from '@components/sections/moreProjects';
-import Youtube from '@components/sections/youtube';
+import AllPortfolioSection from '@components/sections/allPortfolio';
+import MoreProjectsSection from '@components/sections/hobbyProjects';
+import YoutubeShoutout from '@components/sections/youtube';
 
 const ProjectsPage = ({ location }) => (
   <Layout location={location}>
     <Helmet title="Portfolio" />
 
+    <AllPortfolioSection hideViewAllBtn />
 
-    <PortfolioSection hideViewAllBtn />
-
-    <Youtube />
+    <YoutubeShoutout />
 
     <MoreProjectsSection hideViewAllBtn />
   </Layout >
@@ -27,35 +25,3 @@ ProjectsPage.propTypes = {
 
 export default ProjectsPage;
 
-export const pageQuery = graphql`
-  query {
-    projects: allMarkdownRemark(
-      filter: { fileAbsolutePath: { regex: "/portfolio/" } }
-      sort: { fields: [frontmatter___date], order: DESC }
-    ) {
-      edges {
-        node {
-          frontmatter {
-            date
-            description
-            title
-            tech
-            github
-            external
-            company
-            thumbnail {
-              childImageSharp {
-                fluid(quality: 100, maxWidth: 1000) {
-                  ...GatsbyImageSharpFluid
-                }
-              }
-            }
-            slug
-            isFeatured
-          }
-          html
-        }
-      }
-    }
-  }
-`;
